에러처리 다시 봐야함

+++ services/posts.service.js
@@ -4,14 +4,14 @@ class PostService {
     postRepository = new PostRepository();
 
     //게시글 전체 보기
-    findAllPost = async () => {
+    findAllPost = async (userId) => {
         const allPost = await this.postRepository.findAllPost();
+
         const Posts = allPost.posts.map((post, i) => {
             const Locals = allPost.Locals;
             const like = allPost.like;
+            const arr = allPost.arr;
 
-            //위에 map(post,i)가 있어서 굳이 for 문 돌리지 않아도 된다고 함.. 그리고 됨!
-            // for (let i = 0; i <Locals.length; i++) {
             return {
                 postId: post.postId,
                 title: post.title,
@@ -21,13 +21,13 @@ class PostService {
                 MBTI: allPost.Locals[i].dataValues.MBTI,
                 createdAt: post.createdAt,
                 like: allPost.like[i],
+                likepost: allPost.arr[i],
                 commentNum: allPost.CommentNum[i],
                 info: {
                     songTitle: post.songTitle,
                     singer: post.singer,
                 },
             };
-            // }
         });
 
         //게시글 내림 차순으로 정렬
@@ -42,18 +42,25 @@ class PostService {
 
     //게시글 상세조회
     getPost = async (postId) => {
-        const getPostData = await this.postRepository.findOnePost(postId);
-        const getCommentData = await this.postRepository.findComments(postId);
-        const postLikeData = getPostData.dataValues.Likes;
-        const commentUserDate = getPostData.dataValues.Users;
+        if (!postId) {
+            return { success: false };
+        } else {
+            const getPostData = await this.postRepository.findOnePost(postId);
+            const getCommentData = await this.postRepository.findComments(
+                postId
+            );
+            const postLikeData = getPostData.dataValues.Likes;
+            const commentUserDate = getPostData.dataValues.Users;
 
-        delete getPostData.dataValues.Likes; //배열 지우는 방법
+            delete getPostData.dataValues.Likes; //배열 지우는 방법
 
-        return {
-            poster: getPostData,
-            like: postLikeData.length,
-            commenter: getCommentData,
-        };
+            return {
+                success: true,
+                poster: getPostData,
+                like: postLikeData.length,
+                commenter: getCommentData,
+            };
+        }
     };
     //게시글 생성
     createPost = async (
@@ -65,6 +72,11 @@ class PostService {
         userId,
         MBTI
     ) => {
+        //입력 유무 확인
+        if (title || content || imageUrl || songTitle || singer == undefined) {
+            return { success: false };
+        }
+
         await this.postRepository.createPost(
             title,
             content,
@@ -82,18 +94,48 @@ class PostService {
     };
 
     //게시글 수정
-    updatePost = async (postId, title, content, imageUrl) => {
-        await this.postRepository.updatePost(postId, title, content, imageUrl);
-        return {
-            success: true,
-            msg: "게시물이 수정되었습니다.",
-        };
+    updatePost = async (
+        postId,
+        title,
+        content,
+        imageUrl,
+        songTitle,
+        singer
+    ) => {
+        if (
+            !postId ||
+            title ||
+            content ||
+            imageUrl ||
+            songTitle ||
+            singer == undefined
+        ) {
+            return {
+                success: false,
+                msg,
+            };
+        } else {
+            await this.postRepository.updatePost(
+                postId,
+                title,
+                content,
+                imageUrl
+            );
+            return {
+                success: true,
+                msg,
+                // msg: "게시물이 수정되었습니다.",
+            };
+        }
     };
 
     //게시글 삭제
     deletePost = async (postId) => {
+        if (!postId) {
+            return { success: false, msg };
+        }
         await this.postRepository.deletePost(postId);
-        return { success: true, msg: "게시물 삭제에 성공했습니다." };
+        return { success: true, msg };
     };
 }
